+++ rest_framework_json_api/pagination.py
@@ -8,13 +8,24 @@ from rest_framework.pagination import LimitOffsetPagination, PageNumberPaginatio
 from rest_framework.utils.urls import remove_query_param, replace_query_param
 from rest_framework.views import Response
 
+from rest_framework_json_api.settings import json_api_settings
 
-class JsonApiPageNumberPagination(PageNumberPagination):
+
+class PageNumberPagination(PageNumberPagination):
     """
     A json-api compatible pagination format
+
+    An older version of this used `page` and `page_size` so
+    use a setting to enable the "standard" query param names.
     """
-    page_query_param = 'page[number]'
-    page_size_query_param = 'page[size]'
+    if json_api_settings.STANDARD_PAGINATION:
+        page_query_param = 'page[number]'
+        page_size_query_param = 'page[size]'
+    else:
+        page_query_param = 'page'
+        page_size_query_param = 'page_size'
+        warnings.warn("'page' and 'page_size' parameters are deprecated. "
+                      "Set JSON_API_STANDARD_PAGINATION=True", DeprecationWarning)
     max_page_size = 100
 
     def build_link(self, index):
@@ -50,7 +61,7 @@ class JsonApiPageNumberPagination(PageNumberPagination):
         })
 
 
-class JsonApiLimitOffsetPagination(LimitOffsetPagination):
+class LimitOffsetPagination(LimitOffsetPagination):
     """
     A limit/offset based style. For example:
     http://api.example.org/accounts/?page[limit]=100
@@ -58,7 +69,11 @@ class JsonApiLimitOffsetPagination(LimitOffsetPagination):
     """
     limit_query_param = 'page[limit]'
     offset_query_param = 'page[offset]'
-    max_limit = 100
+    if json_api_settings.STANDARD_PAGINATION:
+        max_limit = 100
+    else:
+        warnings.warn("'max_limit = None' is deprecated. "
+                      "Set JSON_API_STANDARD_PAGINATION=True", DeprecationWarning)
 
     def get_last_link(self):
         if self.count == 0:
@@ -100,31 +115,28 @@ class JsonApiLimitOffsetPagination(LimitOffsetPagination):
         })
 
 
-class PageNumberPagination(JsonApiPageNumberPagination):
+class JsonApiPageNumberPagination(PageNumberPagination):
     """
-    Deprecated paginator that uses different query parameters
+    Changed our minds about the naming scheme. Removed JsonApi prefx.
     """
-    page_query_param = 'page'
-    page_size_query_param = 'page_size'
 
     def __init__(self):
         warnings.warn(
-            'PageNumberPagination is deprecated. Use JsonApiPageNumberPagination '
+            'JsonApiPageNumberPagination is deprecated. Use PageNumberPagination '
             'or create custom pagination. See '
             'https://django-rest-framework-json-api.readthedocs.io/en/stable/usage.html#pagination',
             DeprecationWarning)
         super(PageNumberPagination, self).__init__()
 
 
-class LimitOffsetPagination(JsonApiLimitOffsetPagination):
+class JsonApiLimitOffsetPagination(LimitOffsetPagination):
     """
-    Deprecated paginator that uses a different max_limit
+    Changed our minds about the naming scheme. Removed JsonApi prefx.
     """
-    max_limit = None
 
     def __init__(self):
         warnings.warn(
-            'LimitOffsetPagination is deprecated. Use JsonApiLimitOffsetPagination '
+            'JsonApiLimitOffsetPagination is deprecated. Use LimitOffsetPagination '
             'or create custom pagination. See '
             'https://django-rest-framework-json-api.readthedocs.io/en/stable/usage.html#pagination',
             DeprecationWarning)